Mock repositories, unitary tests and code improvements

+++ domain/customer.go
@@ -3,7 +3,7 @@ package domain
 import "context"
 
 type Customer struct {
-	ID        int64  `json:"id"`
+	ID        int    `json:"id"`
 	Firstname string `json:"firstName"`
 	Lastname  string `json:"lastName"`
 	Document  string `json:"document"`
@@ -13,16 +13,16 @@ type Customer struct {
 
 type CustomerUsecase interface {
 	Fetch(ctx context.Context) ([]Customer, error)
-	GetByID(ctx context.Context, id int64) (Customer, error)
+	GetByID(ctx context.Context, id int) (Customer, error)
 	Create(ctx context.Context, customer *Customer) error
 	Update(ctx context.Context, customer *Customer) error
-	Delete(ctx context.Context, id int64) error
+	Delete(ctx context.Context, id int) error
 }
 
 type CustomerRepository interface {
 	Fetch(ctx context.Context) ([]Customer, error)
-	GetByID(ctx context.Context, id int64) (Customer, error)
+	GetByID(ctx context.Context, id int) (Customer, error)
 	Create(ctx context.Context, customer *Customer) error
 	Update(ctx context.Context, customer *Customer) error
-	Delete(ctx context.Context, id int64) error
+	Delete(ctx context.Context, id int) error
 }
